Move logic of program from Program.cs to Lesson3.cs
